feat: add Object.assign for better runtime behavior
Previously, it was the case that an error-like object like could be
provided to the function and NOT have custom fields copied over to the
returned object.

Example error-like object: `{ message: "foo", customField: "bar" }`

With the above example, while the returned error instance would have a
"message" value of "foo", it would not contain "customField" — a result
which is arguably counter-intuitive. Since this change affects the
runtime behavior of a key export, this commit is implemented as a
breaking change.

BREAKING_CHANGE